+++ main.py
@@ -1,40 +1,5 @@
-
-def main():
- for i in range(1, 101):
-    if i % 3 == 0 and i % 5 == 0:
-        print("FizzBuzz")
-    elif i % 3 == 0:
-        print("Fizz")
-    elif i % 5 == 0:
-        print("Buzz")
-    else:
-        print(i)
-
-def fizzbuzz(n):
-    if n % 3 == 0 and n % 5 == 0:
-        return "FizzBuzz"
-    elif n % 3 == 0:
-        return "Fizz"
-    elif n % 5 == 0:
-        return "Buzz"
-    else:
-        return str(n)
-
-for i in range(1, 101):
-    print(fizzbuzz(i))
-
-# Test function
-def test_fizzbuzz():
-    assert fizzbuzz(1) == "1"
-    assert fizzbuzz(3) == "Fizz"
-    assert fizzbuzz(5) == "Buzz"
-    assert fizzbuzz(15) == "FizzBuzz"
-    assert fizzbuzz(30) == "FizzBuzz"
-    assert fizzbuzz(98) == "98"
-    print("All tests passed!")
-
-
+from fizzbuzz import fizzbuzz 
 
 if __name__ == '__main__':
-    main()
-    test_fizzbuzz()
+    for i in range(1, 101):
+        print(fizzbuzz(i))
