+++ test.py
@@ -1,0 +1,14 @@
+from fizzbuzz import fizzbuzz 
+
+
+# Test function
+def test_fizzbuzz():
+    assert fizzbuzz(1) == "1"
+    assert fizzbuzz(3) == "Fizz"
+    assert fizzbuzz(5) == "Buzz"
+    assert fizzbuzz(15) == "FizzBuzz"
+    assert fizzbuzz(30) == "FizzBuzz"
+    assert fizzbuzz(98) == "98"
+    print("All tests passed!") 
+
+test_fizzbuzz()
